requirements.txt
Django==3.1
django-rest-framework
psycopg2-binary